+++ pricing/RELEASES.md
@@ -63,3 +63,11 @@ Based on:
 - Speakeasy CLI 1.376.0 (2.402.5) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.3.0] pricing
+
+## 2024-08-28 00:26:15
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/pricing-api.yaml
+- Speakeasy CLI 1.381.0 (2.404.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.3.1] pricing